Move list_import tests to their own file
